fix: large files broke prefetch
backport master commit 7e39a5d8f1a7dc3555ce474d5e6b0c76c5df7e79
Author: 泰友 <cuichengxu.ccx@antgroup.com>
Date:   Wed Jul 5 14:14:09 2023 +0800

    fix: large files broke prefetch

    Files larger than 4G leads to prefetch panic, because the max blob io
    range is smaller than 4G. This pr changes blob io max size from u32 to
    u64.

+++ rafs/src/fs.rs
@@ -797,7 +797,7 @@ impl FileSystem for Rafs {
             let r = self.device.read_to(w, desc)?;
             result += r;
             recorder.mark_success(r);
-            if r as u32 != desc.bi_size {
+            if r as u64 != desc.bi_size {
                 break;
             }
         }
